+++ TuCam/Ressources/Assets.xcassets/AppIcon.appiconset/Contents.json
