+++ TuCam/Source/HomeViewController.swift
@@ -1,0 +1,26 @@
+//
+//  HomeViewController.swift
+//  TuCam
+//
+//  Created by Natalie Zhang on 2021-02-08.
+//
+
+import UIKit
+
+class HomeViewController: UIViewController {
+	init() {
+		super.init(nibName: nil, bundle: nil)
+		title = "Home"
+	}
+	
+	required init?(coder: NSCoder) {
+		fatalError("init(coder:) has not been implemented")
+	}
+	
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+
+}
+
