--- TuCam/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  TuCam
-//
-//  Created by Natalie Zhang on 2021-02-08.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
